requirements.txt
APScheduler==3.9.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
click==8.0.4
colorama==0.4.4
colorhash==1.0.4
configparser==5.2.0
cycler==0.11.0
Flask==2.0.3
Flask-Cors==3.0.8
Flask-MonitoringDashboard==3.0.6
fonttools==4.30.0
greenlet==1.1.2
imbalanced-learn==0.6.1
imblearn==0.0
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
kneed==0.5.1
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
psutil==5.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
scikit-learn==0.22.1
scipy==1.7.3
six==1.16.0
sklearn-pandas==1.8.0
SQLAlchemy==1.4.32
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
Werkzeug==2.0.3
wincertstore==0.2
xgboost==1.5.2
zipp==3.7.0
